refactor(group): apply rabbit's advice

+++ src/main/java/sleppynavigators/studyupbackend/presentation/group/dto/response/GroupMemberListResponse.java
@@ -43,7 +43,7 @@ public record GroupMemberListResponse(
                     groupMember.getUser().getUserProfile().getUsername(),
                     groupMember.getUser().getPoint().getAmount(),
                     groupMember.calcAvgChallengeCompletionRate(),
-                    groupMember.calcHuntingCount().longValue()
+                    groupMember.calcHuntingCount()
             );
         }
     }
